keos_project4::round_robin

Struct RoundRobin

pub struct RoundRobin {
    percores: [PerCore; 4],
}

A round robin scheduler.

The RoundRobin struct represents a round-robin scheduler, which is a type of preemptive scheduling used in operating systems to manage the execution of processes. In a round-robin scheduler, each process is assigned a fixed time slice (quantum) in which it can run. Once the time slice expires, the process is suspended, and the next process in the queue is scheduled to run. This continues in a circular manner, ensuring fair distribution of CPU time among processes.

Fields
percores: [PerCore; 4]
Implementations

impl RoundRobin

pub fn new() -> Self

Create a new RoundRobin scheduler.